requirements.txt
Django
torch
torchvision
numpy
opencv-python
pillow
deep_utils
matplotlib
gdown
docker-compose
termcolor
facenet-pytorch